Changeset formatting and sonar fixes

+++ hoot-core/src/main/cpp/hoot/core/algorithms/changeset/ChangesetDeriver.h
@@ -22,7 +22,7 @@
  * This will properly maintain the copyright information. Maxar
  * copyrights will be updated automatically.
  *
- * @copyright Copyright (C) 2015, 2016, 2017, 2018, 2019, 2020, 2021 Maxar (http://www.maxar.com/)
+ * @copyright Copyright (C) 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022 Maxar (http://www.maxar.com/)
  */
 #ifndef CHANGESETDERIVER_H
 #define CHANGESETDERIVER_H
@@ -69,8 +69,8 @@ public:
    */
   Change readNextChange() override;
 
-  int getNumFromElementsParsed() const override { return _numFromElementsParsed; }
-  int getNumToElementsParsed() const override { return _numToElementsParsed; }
+  int getNumFromElementsParsed() const override { return static_cast<int>(_numFromElementsParsed); }
+  int getNumToElementsParsed() const override { return static_cast<int>(_numToElementsParsed); }
 
   int getNumCreateChanges() const override
   { return _changesByType[Change::ChangeType::Create]; }
